Use Static data

+++ src/Financier.Desktop/Helpers/MapperHelper.cs
@@ -9,14 +9,14 @@ namespace Financier.Desktop.Helpers
     {
         public static void MapTransfer(TransferDto dto, Transaction tr)
         {
-            tr.FromAccountId = dto.FromAccountId;
-            tr.ToAccountId = dto.ToAccountId;
+            tr.FromAccountId = (int)dto.FromAccountId;
+            tr.ToAccountId = (int)dto.ToAccountId;
             tr.Note = dto.Note;
             tr.FromAmount = Math.Abs(dto.FromAmount) * -1;
             tr.ToAmount = Math.Abs(dto.ToAmount == 0 ? dto.FromAmount : dto.ToAmount);
             tr.DateTime = UnixTimeConverter.ConvertBack(dto.DateTime);
             tr.LastRecurrence = UnixTimeConverter.ConvertBack(DateTime.Now);
-            tr.OriginalCurrencyId = dto.FromAccount.CurrencyId;
+            tr.OriginalCurrencyId = (int)dto.FromAccount.CurrencyId;
             tr.OriginalFromAmount = Math.Abs(dto.FromAmount) * -1;
             tr.CategoryId = 0;
             tr.Category = default;
@@ -24,7 +24,7 @@ namespace Financier.Desktop.Helpers
 
         public static void MapTransaction(TransactionDto dto, Transaction tr)
         {
-            tr.FromAccountId = dto.AccountId;
+            tr.FromAccountId = (int)dto.AccountId;
 
             if (dto.OriginalCurrencyId > 0 && dto.Account?.CurrencyId == dto.OriginalCurrencyId)
             {
@@ -36,19 +36,19 @@ namespace Financier.Desktop.Helpers
             {
                 tr.FromAmount = Math.Abs(dto.FromAmount) * (dto.IsAmountNegative ? -1 : 1);
                 tr.OriginalFromAmount = dto.OriginalFromAmount ?? 0;
-                tr.OriginalCurrencyId = dto.OriginalCurrencyId ?? 0;
+                tr.OriginalCurrencyId = (int)(dto.OriginalCurrencyId ?? 0);
             }
 
-            tr.CategoryId = dto.CategoryId ?? 0;
+            tr.CategoryId = (int)(dto.CategoryId ?? 0);
             tr.Category = default;
             tr.Location = default;
             tr.Project = default;
             tr.OriginalCurrency = default;
             tr.FromAccount = default;
             tr.ToAccount = default;
-            tr.PayeeId = dto.PayeeId ?? 0;
-            tr.LocationId = dto.LocationId ?? 0;
-            tr.ProjectId = dto.CategoryId == Category.Split.Id ? 0 : (dto.ProjectId ?? 0);
+            tr.PayeeId = (int)(dto.PayeeId ?? 0);
+            tr.LocationId = (int)(dto.LocationId ?? 0);
+            tr.ProjectId = (int)(dto.CategoryId == Category.Split.Id ? 0 : (dto.ProjectId ?? 0));
             tr.Note = dto.Note;
             tr.DateTime = UnixTimeConverter.ConvertBack(dto.DateTime);
             tr.LastRecurrence = UnixTimeConverter.ConvertBack(DateTime.Now);
